refactor(dashboard): extract Court sessions + audit + rfq sub-routers (F-B-202 PR-9/10)
Audit ref: imp/audits/2026-05-20/findings/track-b/F-B-202.md

Sprint 2 modularization PR-9 of 10. base=main.

Extracts 6 routes (1 sessions + 2 audit + 3 rfq) from `app/dashboard/router.py`
into three new sibling modules. Pure cut+paste; behaviour-equivalent.

  app/dashboard/sessions_routes.py (new, 84 LOC, 1 route):
    GET  /dashboard/sessions                  session list + status filter

  app/dashboard/audit_routes.py (new, 108 LOC, 2 routes):
    GET  /dashboard/audit                     log table + LIKE search
    POST /dashboard/audit/verify              hash chain integrity check (admin)

  app/dashboard/rfq_routes.py (new, 155 LOC, 3 routes):
    GET  /dashboard/rfqs                      rfq list (last 100)
    GET  /dashboard/rfq/{rfq_id}              detail + responses
    POST /dashboard/rfq/{rfq_id}/approve      approve quote + mint txn token

  app/dashboard/router.py (2071 → 1840 LOC, -231 net):
    - Sections 13 (Sessions) + 14 (Audit Log) + 15 (RFQ) removed;
      breadcrumb comments at the old positions.
    - Three new include_router calls.
    - Unused imports trimmed: Query (fastapi); or_ (sqlalchemy);
      SessionMessageRecord + RfqRecord + RfqResponseRecord (broker.db_models);
      create_transaction_token + compute_payload_hash (auth.transaction_token,
      whole module).

Behavioural delta: zero. Admin-only + CSRF gate on /audit/verify
preserved. RFQ approve still mints transaction token + pushes WS event
on the same path as before (admin path).

Pattern follows orgs_routes.py / auth_routes.py: sub-router declares
``APIRouter(tags=[...])`` with **no prefix**, route paths bare; the
outer ``/dashboard`` prefix is prepended by ``router.include_router(...)``
in router.py.

Tests:
  - pytest tests/test_dashboard.py tests/test_audit_r2_t1.py
    tests/test_dashboard_principals.py — 73/73
  - pytest tests/test_dashboard*.py tests/test_audit*.py tests/test_rfq*.py
    tests/test_onboarding.py — 421/421 (1 pre-existing env-fail deselected:
    test_badge_approvals_empty_when_plugin_unavailable requires
    cullis_enterprise NOT installed; passes clean in CI).

Router LOC progress:
  Pre-Sprint 2 : 67 routes, 3125 LOC
  Post-PR-1 #839: 67 routes, 2930 LOC   merged
  Post-PR-2 #841: 60 routes, 2614 LOC   merged
  Post-PR-3 #842: 57 routes, 2528 LOC   merged
  Post-PR-4 #847: 50 routes, 2372 LOC   merged
  Post-PR-5 #848: 39 routes, 2071 LOC   merged
  Post-PR-9 (this, before PR-6/7/8 merged): 33 routes, 1840 LOC
  Target post-PR-10: ~3 (aggregator), ~100 LOC

Conflict warning: PR-6 #849 + PR-7 #850 + PR-8 #851 are still open.
This PR's sections (13+14+15) have zero line overlap with their
sections — rebase clean post-merge in any order. Only contention is
the include_router stack at top of router.py (1-line nest, trivial
to resolve).

Next PR-10 (closing F-B-202): agents_demo sub-router (sections 11+12)
+ final aggregator pass. Target ~3 routes, ~100 LOC end-state.
